requirements.txt
# Required Waggle Python modules
https://github.com/waggle-sensor/pywaggle/archive/refs/tags/v0.40.4.zip

# Custom Python modules. Please add your requirements below.
librosa==0.8.0
numpy==1.19.5
pillow==8.1.0
tensorflow==2.5.0                 
